test: ユニットテストファイルを分割して整理
「タイトル」ボタンのテストを、別ファイルに分割した。
