+++ src/app/components/BookmarkManager.test.tsx
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import fetchMock from "jest-fetch-mock";
 import React, { act } from "react";
 
-import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 
 import { MessageProvider } from "../contexts/MessageContext";
 import BmMessage from "./bmMessage";
@@ -84,192 +84,6 @@ describe("BookmarkManagerの表示を確認", () => {
       expect(screen.getByTestId("bm-message")).toHaveTextContent(
         /Failed to fetch: \[500\] Internal Server Error$/
       );
-    });
-  });
-});
-
-describe("BookmarkManagerのURLとタイトルのテキストとボタンのテスト", () => {
-  beforeEach(() => {
-    fetchMock.resetMocks();
-  });
-
-  it("タイトルを取得するボタンをクリック", async () => {
-    // タイトルを取得するボタンをクリックすると、タイトルを取得するAPIを呼び出す。
-    // パラメータとしてURLのテキストボックスに入力された文字列が渡される。
-    // タイトルのテキストボックスに、APIから返されたタイトルが表示される。
-
-    const url = "https://mail.google.com/mail/";
-    const title = "Gmail";
-
-    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
-
-    await act(async () => {
-      render(
-        <MessageProvider>
-          <BmMessage />
-          <BookmarkManager />
-        </MessageProvider>
-      );
-    });
-
-    fetchMock.resetMocks();
-
-    const urlInput = screen.getByRole("textbox", { name: "url" });
-    const titleInput = screen.getByRole("textbox", { name: "title" });
-
-    const titleButton = screen.getByRole("button", { name: "タイトル" });
-
-    // タイトルをAPIで取得する前は、クリアされていることを確認
-    await act(async () => {
-      fireEvent.change(titleInput, { target: { value: "" } });
-    });
-    await waitFor(() => {
-      expect(titleInput).toHaveValue("");
-    });
-
-    fetchMock.mockResponseOnce(title);
-
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: url } });
-      fireEvent.click(titleButton);
-    });
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledTimes(1);
-      expect(fetchMock.mock.calls[0][0]).toEqual("/api/title?url=" + url);
-      expect(titleInput).toHaveValue(title);
-    });
-  });
-
-  it("パラメータとして渡されたURLにアクセスできない場合、メッセージ領域にエラーメッセージを表示する。", async () => {
-    // タイトルを取得するボタンをクリックして、エラーコード500の場合、タイトルのテキストボックスに、なにも表示しない。
-    const url = "https://mail.google.com/mail/";
-
-    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
-
-    // render(
-    //   <MessageProvider>
-    //     <BmMessage />
-    //     <BookmarkManager />
-    //   </MessageProvider>
-    // );
-    await act(async () => {
-      render(
-        <MessageProvider>
-          <BmMessage />
-          <BookmarkManager />
-        </MessageProvider>
-      );
-    });
-
-    fetchMock.resetMocks();
-
-    fetchMock.mockResponseOnce("Can't find title", {
-      status: 500,
-      headers: { "Content-Type": "text/plain" },
-    });
-
-    const urlInput = screen.getByRole("textbox", { name: "url" });
-    const titleInput = screen.getByRole("textbox", { name: "title" });
-
-    const titleButton = screen.getByRole("button", { name: "タイトル" });
-
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: url } });
-      fireEvent.click(titleButton);
-    });
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledTimes(1);
-      expect(fetchMock.mock.calls[0][0]).toEqual("/api/title?url=" + url);
-
-      expect(screen.getByTestId("bm-message")).toHaveTextContent(
-        "Can't find title: [500] " + url
-      );
-      expect(screen.getByRole("button", { name: "確認" })).toBeInTheDocument();
-      expect(titleInput).toHaveValue("");
-    });
-  });
-
-  it("APIからタイトルを取得した後に、テキストボックスでタイトルを編集した場合、更新ボタンをクリックすると編集後のテキストが渡される。", async () => {
-    // APIからタイトルを取得した後に、テキストボックスでタイトルを編集した場合、更新ボタンをクリックすると編集後のテキストが渡される。
-    const url = "https://mail.google.com/mail/";
-    const title = "Gmail";
-    const title_edited = "GMAIL";
-
-    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
-
-    await act(async () => {
-      render(
-        <MessageProvider>
-          <BmMessage />
-          <BookmarkManager />
-        </MessageProvider>
-      );
-    });
-
-    fetchMock.resetMocks();
-
-    const urlInput = screen.getByRole("textbox", { name: "url" });
-    const titleInput = screen.getByRole("textbox", { name: "title" });
-
-    const titleButton = screen.getByRole("button", { name: "タイトル" });
-    const updateButton = screen.getByRole("button", { name: "追加" });
-
-    fetchMock.mockResponseOnce(title);
-
-    fireEvent.change(urlInput, { target: { value: url } });
-    fireEvent.click(titleButton);
-
-    fireEvent.change(titleInput, { target: { value: "" } });
-    fireEvent.change(titleInput, { target: { value: title_edited } });
-    fireEvent.click(updateButton);
-
-    await waitFor(() => {
-      expect(titleInput).toHaveValue(title_edited);
-      expect(fetchMock).toHaveBeenCalledTimes(2);
-      expect(fetchMock.mock.calls[0][0]).toEqual("/api/title?url=" + url);
-      expect(fetchMock.mock.calls[1][0]).toEqual("/api/bookmark");
-    });
-  });
-
-  it("APIからタイトルが返ってこない場合、メッセージ領域にエラーメッセージを表示する。", async () => {
-    // タイトルを取得するボタンをクリックして、エラーコード500の場合、タイトルのテキストボックスに、なにも表示しない。
-    const url = "https://mail.google.com/mail/";
-
-    fetchMock.mockResponseOnce(JSON.stringify(mockBookmarks));
-
-    await act(async () => {
-      render(
-        <MessageProvider>
-          <BmMessage />
-          <BookmarkManager />
-        </MessageProvider>
-      );
-    });
-
-    fetchMock.resetMocks();
-    fetchMock.mockResponseOnce("");
-
-    const urlInput = screen.getByRole("textbox", { name: "url" });
-    const titleInput = screen.getByRole("textbox", { name: "title" });
-
-    const titleButton = screen.getByRole("button", { name: "タイトル" });
-
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: url } });
-      fireEvent.click(titleButton);
-    });
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledTimes(1);
-      expect(fetchMock.mock.calls[0][0]).toEqual("/api/title?url=" + url);
-
-      expect(screen.getByTestId("bm-message")).toHaveTextContent(
-        "Can't find title: " + url
-      );
-      expect(screen.getByRole("button", { name: "確認" })).toBeInTheDocument();
-      expect(titleInput).toHaveValue("");
     });
   });
 });
